+++ server/model/api.go
@@ -1,18 +1,10 @@
-package api
+package model
 
 import (
 	"time"
-	"traQer/api/db"
-	"traQer/api/qapi"
 
 	"github.com/traPtitech/go-traq"
 )
-
-type ApiHandler struct {
-	db         *db.DBHandler     // db関連
-	qapi       *qapi.QapiHandler // traqApi関連
-	serverData ServerData // サーバの持つデータ
-}
 
 type ServerData struct {
 	lastTrackMessage traq.Message // 最後に取得したメッセージ
