+++ server/model/bot.go
@@ -1,0 +1,3 @@
+package model
+
+
