+++ server/model/qapi.go
@@ -1,0 +1,8 @@
+package model
+
+import "github.com/traPtitech/go-traq"
+
+type QUserMessages struct {
+	User              traq.User
+	TotalMessageCount int64
+}
