+++ server/service/channel/handler.go
@@ -2,6 +2,10 @@ package channel
 
 import "traQer/api"
 
+type ChannelHandler struct {
+	api *api.ApiHandler
+}
+
 func NewChannelHandler(api *api.ApiHandler) *ChannelHandler {
 	return &ChannelHandler{api: api}
 }
