--- server/service/channel/model.go
@@ -1,7 +1,0 @@
-package channel
-
-import "traQer/api"
-
-type ChannelHandler struct {
-	api *api.ApiHandler
-}
