--- server/service/model.go
@@ -1,11 +1,0 @@
-package service
-
-import (
-	"traQer/service/channel"
-	"traQer/service/user"
-)
-
-type Service struct {
-	user    *user.UserHandler
-	channel *channel.ChannelHandler
-}
