+++ server/service/user/handler.go
@@ -2,6 +2,11 @@ package user
 
 import "traQer/api"
 
+// Handler
+type UserHandler struct {
+	api *api.ApiHandler
+}
+
 func NewUserHandler(api *api.ApiHandler) *UserHandler {
 	return &UserHandler{api: api}
 }
